refator: 앨범 상세 공용 컴포넌트 분리 #149
