+++ VibeTrip/Features/Album/Views/AlbumDetailView.swift
@@ -879,237 +879,6 @@ private extension AlbumDetailView {
     }
 }
 
-// MARK: - AlbumDetailNavigationOverlay
-
-private struct AlbumDetailNavigationOverlay: View {
-    
-    private enum Constants {
-        static let horizontalPadding: CGFloat = 20
-        static let topPadding: CGFloat = 12
-        static let iconSize: CGFloat = 22
-        static let touchTargetSize: CGFloat = 44
-    }
-    
-    let onBackTap: () -> Void
-    let onMoreTap: () -> Void
-    
-    var body: some View {
-        HStack {
-            // 뒤로가기 버튼
-            Button(action: onBackTap) {
-                Image(systemName: "chevron.left")
-                    .font(.system(size: Constants.iconSize, weight: .semibold))
-                    .foregroundStyle(.white)
-                    .blendMode(.difference)
-                    .frame(width: Constants.touchTargetSize, height: Constants.touchTargetSize)
-            }
-            
-            Spacer()
-            
-            // 앨범 옵션 버튼
-            Button(action: onMoreTap) {
-                Image(systemName: "ellipsis")
-                    .font(.system(size: Constants.iconSize, weight: .semibold))
-                    .foregroundStyle(.white)
-                    .blendMode(.difference)
-                    .frame(width: Constants.touchTargetSize, height: Constants.touchTargetSize)
-            }
-        }
-        .compositingGroup()
-        .padding(.horizontal, Constants.horizontalPadding)
-        .padding(.top, Constants.topPadding)
-        .frame(maxWidth: .infinity, alignment: .top)
-    }
-}
-
-// MARK: - AlbumDetailActionButton
-
-private struct AlbumDetailActionButton: View {
-    
-    private enum Constants {
-        static let height: CGFloat = 48
-        static let cornerRadius: CGFloat = 24
-        static let iconTextSpacing: CGFloat = 4
-        static let fontSize: CGFloat = 14
-        static let iconSize: CGFloat = 18
-        static let assetIconSize: CGFloat = 22
-        
-        static let iconFrameWidth: CGFloat = 22
-        static let horizontalPadding: CGFloat = 20
-        static let verticalPadding: CGFloat = 12
-    }
-    
-    let title: String
-    let systemImageName: String
-    var isAssetImage: Bool = false
-    var iconColor: Color = Color("appPrimary/400")
-    var referenceTitle: String? = nil
-    let action: () -> Void
-    
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: Constants.iconTextSpacing) {
-                // 아이콘 고정 너비
-                ZStack(alignment: .topTrailing) {
-                    Group {
-                        if isAssetImage {
-                            Image(systemImageName)
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: Constants.assetIconSize, height: Constants.assetIconSize)
-                        } else {
-                            Image(systemName: systemImageName)
-                                .font(.system(size: Constants.iconSize, weight: .medium))
-                                .contentTransition(.symbolEffect(.replace.offUp)) /// 심볼 전환 효과
-                                .foregroundStyle(iconColor)
-                        }
-                    }
-                    .frame(width: Constants.iconFrameWidth, height: Constants.assetIconSize)
-                }
-                
-                ZStack {
-                    if let ref = referenceTitle {
-                        Text(ref)
-                            .opacity(0) // 레이아웃 너비 고정용
-                    }
-                    Text(title)
-                        .fixedSize()
-                        .id(title)
-                        .transition(.asymmetric(    /// 타이틀 전환 효과
-                            insertion: .opacity.animation(.easeIn(duration: 0.2)),
-                            removal: .opacity.animation(.easeOut(duration: 0.05))
-                                               ))
-                }
-                .font(.setPretendard(weight: .medium, size: Constants.fontSize))
-            }
-            .foregroundStyle(Color("appPrimary/500"))
-            .padding(.horizontal, Constants.horizontalPadding)
-            .padding(.vertical, Constants.verticalPadding)
-            .frame(maxWidth: .infinity, minHeight: Constants.height)
-            .background(Color("appPrimary/400").opacity(0.12))
-            .clipShape(RoundedRectangle(cornerRadius: Constants.cornerRadius))
-        }
-        .buttonStyle(.plain)
-    }
-}
-
-// MARK: - AlbumDetailEmptyStateSection
-// 표시할 로그 없을 시 안내 문구
-
-private struct AlbumDetailEmptyStateSection: View {
-    
-    private enum Constants {
-        static let spacing: CGFloat = 8
-        static let titleFontSize: CGFloat = 16
-        static let descriptionFontSize: CGFloat = 14
-    }
-    
-    var body: some View {
-        VStack(spacing: Constants.spacing) {
-            Text("아직 기록된 로그가 없어요.")
-                .font(.setPretendard(weight: .semiBold, size: Constants.titleFontSize))
-                .foregroundStyle(Color("GrayScale/500"))
-                .multilineTextAlignment(.center)
-            
-            Text("로그를 작성하고 여행의 추억을 완성해 보세요.")
-                .font(.setPretendard(weight: .medium, size: Constants.descriptionFontSize))
-                .foregroundStyle(Color("GrayScale/400"))
-                .multilineTextAlignment(.center)
-        }
-        .frame(maxWidth: .infinity)
-    }
-}
-
-// MARK: - AlbumMenuItemButtonStyle
-
-private struct AlbumMenuItemButtonStyle: ButtonStyle {
-    
-    private enum Constants {
-        static let horizontalPadding: CGFloat = 16
-        static let verticalPadding: CGFloat = 8
-        static let cornerRadius: CGFloat = 8
-        static let highlightBackground = Color(red: 0.92, green: 0.92, blue: 0.98)
-    }
-    
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .padding(.horizontal, Constants.horizontalPadding)
-            .padding(.vertical, Constants.verticalPadding)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(configuration.isPressed ? Constants.highlightBackground : Color.clear)
-            .cornerRadius(Constants.cornerRadius)
-            .animation(.easeInOut(duration: 0.1), value: configuration.isPressed)
-    }
-}
-
-// MARK: - AlbumDetailAlbumMenuPopup
-
-private struct AlbumDetailAlbumMenuPopup: View {
-    
-    private enum Constants {
-        static let popupWidth: CGFloat = 160
-        static let padding: CGFloat = 8
-        static let itemSpacing: CGFloat = 8
-        static let cornerRadius: CGFloat = 12
-        static let itemFontSize: CGFloat = 14
-    }
-    
-    let isMusicUrlReady: Bool
-    let onDownloadMusic: () -> Void
-    let onEditAlbum: () -> Void
-    let onDeleteAlbum: () -> Void
-    let onReport: () -> Void
-    
-    // 팝업 메뉴 항목
-    private enum MenuItem: CaseIterable {
-        case downloadMusic, editAlbum, deleteAlbum, report
-        
-        var title: String {
-            switch self {
-            case .downloadMusic: return "배경 음악 다운로드"
-            case .editAlbum:    return "앨범 수정"
-            case .deleteAlbum:  return "앨범 삭제"
-            case .report:       return "AI 음악 신고하기"
-            }
-        }
-    }
-    
-    var body: some View {
-        VStack(alignment: .center, spacing: Constants.itemSpacing) {
-            ForEach(MenuItem.allCases, id: \.self) { item in
-                menuItem(item)
-            }
-        }
-        .padding(Constants.padding)
-        .frame(width: Constants.popupWidth, alignment: .center)
-        .background(.white)
-        .cornerRadius(Constants.cornerRadius)
-        // 앨범메뉴팝업 -> detailMenu shadow
-        .appShadow(.detailMenu)
-    }
-    
-    @ViewBuilder
-    private func menuItem(_ item: MenuItem) -> some View {
-        let isDisabled = item == .downloadMusic && !isMusicUrlReady
-        Button {
-            switch item {
-            case .downloadMusic: onDownloadMusic()
-            case .editAlbum:     onEditAlbum()
-            case .deleteAlbum:   onDeleteAlbum()
-            case .report:        onReport()
-            }
-        } label: {
-            HStack(alignment: .center, spacing: 0) {
-                Text(item.title)
-                    .font(.setPretendard(weight: .medium, size: Constants.itemFontSize))
-                    .foregroundStyle(isDisabled ? Color("GrayScale/50") : Color.text)
-            }
-        }
-        .disabled(isDisabled)
-        .buttonStyle(AlbumMenuItemButtonStyle())
-    }
-}
-
 // MARK: - Preview
 
 #Preview("로그 없음") {
